Added Charts.

+++ LocalParks/LocalParks.Core/Chart/ChartDataset.cs
@@ -3,7 +3,7 @@
     public class Dataset
     {
         public string Label { get; set; }
-        public int[] Data { get; set; }
+        public decimal[] Data { get; set; }
         public string[] BackgroundColor { get; set; }
         public string[] BorderColor { get; set; }
         public int BorderWidth { get; set; }
